refactor: #19 filter, resolver 패키지 수정
