+++ src/main/java/site/dogether/auth/controller/AuthController.java
@@ -14,8 +14,8 @@ import org.springframework.web.bind.annotation.RestController;
 import site.dogether.auth.controller.request.LoginRequest;
 import site.dogether.auth.controller.request.WithdrawRequest;
 import site.dogether.auth.controller.response.LoginResponse;
-import site.dogether.auth.resolver.Authentication;
-import site.dogether.auth.service.JwtToken;
+import site.dogether.common.config.web.resolver.Authentication;
+import site.dogether.auth.infrastructure.JwtToken;
 import site.dogether.common.controller.response.ApiResponse;
 import site.dogether.member.service.MemberService;
 import site.dogether.member.service.dto.AuthenticatedMember;
