+++ src/main/java/site/dogether/auth/infrastructure/JwtHandler.java
@@ -1,4 +1,4 @@
-package site.dogether.auth.service;
+package site.dogether.auth.infrastructure;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
