+++ src/main/java/site/dogether/common/config/web/resolver/AuthenticationArgumentResolver.java
@@ -1,4 +1,4 @@
-package site.dogether.auth.resolver;
+package site.dogether.common.config.web.resolver;
 
 import org.springframework.core.MethodParameter;
 import org.springframework.stereotype.Component;
@@ -6,7 +6,7 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-import site.dogether.auth.service.JwtToken;
+import site.dogether.auth.infrastructure.JwtToken;
 
 @Component
 public class AuthenticationArgumentResolver implements HandlerMethodArgumentResolver {
