+++ src/main/java/site/dogether/member/service/MemberService.java
@@ -5,8 +5,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import site.dogether.auth.controller.request.LoginRequest;
 import site.dogether.auth.controller.request.WithdrawRequest;
-import site.dogether.auth.service.JwtHandler;
-import site.dogether.auth.service.JwtToken;
+import site.dogether.auth.infrastructure.JwtHandler;
+import site.dogether.auth.infrastructure.JwtToken;
 import site.dogether.member.domain.Member;
 import site.dogether.member.infrastructure.entity.MemberJpaEntity;
 import site.dogether.member.infrastructure.repository.MemberJpaRepository;
